feat: 중기청 상품 추가

+++ src/main/java/com/bankersCalculator/server/advice/loanAdvice/dto/request/LoanAdviceRequest.java
@@ -58,7 +58,7 @@ public class LoanAdviceRequest {
     // 주택정보
     private RentHousingType rentHousingType;  // 주택타입
 
-    private Double exclusiveArea; // 전용면적
+    private BigDecimal exclusiveArea; // 전용면적
 
     private String buildingName;   // 건물명
