+++ src/main/java/com/bankersCalculator/server/advice/userInputInfo/dto/UserInputInfoResponse.java
@@ -7,18 +7,20 @@ import com.bankersCalculator.server.housingInfo.rentTransactionInquiry.common.Re
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 public class UserInputInfoResponse {
 
     // 고객 정보
-    private Long rentalDeposit; // 임차보증금
-    private Long monthlyRent;   // 월세
-    private Long cashOnHand;    // 보유현금
+    private BigDecimal rentalDeposit; // 임차보증금
+    private BigDecimal monthlyRent;   // 월세
+    private BigDecimal cashOnHand;    // 보유현금
     private Integer age;    // 만나이
     private MaritalStatus maritalStatus; // 혼인상태
-    private Long annualIncome;  //연소득
-    private Long spouseAnnualIncome;    // 배우자연소득
+    private BigDecimal annualIncome;  //연소득
+    private BigDecimal spouseAnnualIncome;    // 배우자연소득
     private ChildStatus childStatus;    // 자녀상태
     private Boolean hasNewborn; // 신생아여부
     private Boolean isSMEEmployee; // 중소기업재직여부(SME: SmallMediumEnterprise);
@@ -27,7 +29,7 @@ public class UserInputInfoResponse {
 
     // 주택정보
     private RentHousingType rentHousingType;  // 주택타입
-    private Double exclusiveArea; // 전용면적
+    private BigDecimal exclusiveArea; // 전용면적
     private String buildingName;   // 건물명
     private String districtCode; // 법정동 코드
     private String dongName;   // 읍명동이름
